+++ src/app/[variants]/(main)/chat/_layout/Desktop/index.tsx
@@ -2,7 +2,6 @@ import { Suspense } from 'react';
 import { Flexbox } from 'react-layout-kit';
 
 import { isDesktop } from '@/const/version';
-import InitClientDB from '@/features/InitClientDB';
 import ProtocolUrlHandler from '@/features/ProtocolUrlHandler';
 
 import { LayoutProps } from '../type';
@@ -22,7 +21,6 @@ const Layout = ({ children, session }: LayoutProps) => {
         <SessionPanel>{session}</SessionPanel>
         <Workspace>{children}</Workspace>
       </Flexbox>
-      {!isDesktop && <InitClientDB bottom={60} />}
       {/* ↓ cloud slot ↓ */}
 
       {/* ↑ cloud slot ↑ */}