+++ src/app/[variants]/loading/index.tsx
@@ -1,10 +1,1 @@
-import { isServerMode } from '@/const/version';
-
-import Client from './Client';
-import Server from './Server';
-
-const ScreenLoading = () => (isServerMode ? <Server /> : <Client />);
-
-ScreenLoading.displayName = 'ScreenLoading';
-
-export default ScreenLoading;
+export { default } from './Server';