+++ src/libs/next-auth/sso-providers/keycloak.ts
@@ -8,9 +8,6 @@ const provider = {
     ...CommonProviderConfig,
     // Specify auth scope, at least include 'openid email'
     authorization: { params: { scope: 'openid email profile' } },
-    clientId: process.env.AUTH_KEYCLOAK_ID,
-    clientSecret: process.env.AUTH_KEYCLOAK_SECRET,
-    issuer: process.env.AUTH_KEYCLOAK_ISSUER,
     profile(profile) {
       return {
         email: profile.email,