+++ src/libs/trpc/client/index.ts
@@ -1,5 +1,4 @@
 export { asyncClient } from './async';
 export * from './desktop';
-export { edgeClient } from './edge';
 export * from './lambda';
 export * from './tools';