+++ src/tools/index.ts
@@ -4,7 +4,6 @@ import { isDesktop } from '@/const/version';
 
 import { ArtifactsManifest } from './artifacts';
 import { CodeInterpreterManifest } from './code-interpreter';
-import { DalleManifest } from './dalle';
 import { LocalSystemManifest } from './local-system';
 import { WebBrowsingManifest } from './web-browsing';
 
@@ -12,11 +11,6 @@ export const builtinTools: LobeBuiltinTool[] = [
   {
     identifier: ArtifactsManifest.identifier,
     manifest: ArtifactsManifest,
-    type: 'builtin',
-  },
-  {
-    identifier: DalleManifest.identifier,
-    manifest: DalleManifest,
     type: 'builtin',
   },
   {